五部曲删除修改

+++ sales/protected/commands/FivestepCommand.php
@@ -7,15 +7,17 @@ class FivestepCommand extends CConsoleCommand
         $sql="select a.id ,a.filename from sales$suffix.sal_fivestep a 
 	          left outer join security$suffix.sec_user b on a.username=b.username
 	          where b.status='I'";
-        $arr=Yii::app()->db->createCommand($sql)->queryRow();
-        foreach ($arr as $a){
-            if(file_exists($a['filename'])){
-                unlink($a['filename']);
+        $arr=Yii::app()->db->createCommand($sql)->queryAll();
+        if(!empty($arr)){
+            foreach ($arr as $a){
+                if(file_exists($a['filename'])){
+                    unlink($a['filename']);
+                }
+                $sql1="delete from sales$suffix.sal_fivestep where id = '".$a['id']."'";
+                $command=Yii::app()->db->createCommand($sql1)->execute();
+                $sql2="delete from sales$suffix.sal_fivestep_info where five_id = '".$a['id']."'";
+                $commands=Yii::app()->db->createCommand($sql2)->execute();
             }
-            $sql1="delete from sales$suffix.sal_fivestep where id = '".$a['id']."'";
-            $command=Yii::app()->db->createCommand($sql1)->execute();
-            $sql2="delete from sales$suffix.sal_fivestep_info where five_id = '".$a['id']."'";
-            $commands=Yii::app()->db->createCommand($sql2)->execute();
         }
     }
 }
